+++ app/uk/gov/hmrc/plasticpackagingtaxregistration/connectors/models/eis/subscription/ChangeOfCircumstanceDetails.scala
@@ -14,20 +14,18 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.plasticpackagingtaxregistration.models
+package uk.gov.hmrc.plasticpackagingtaxregistration.connectors.models.eis.subscription
 
-import play.api.libs.json.{Format, Json}
+import play.api.libs.json.{Json, OFormat}
 
-case class IncorporationDetails(
-  companyNumber: String,
-  companyName: String,
-  ctutr: String,
-  businessVerificationStatus: String,
-  companyAddress: IncorporationAddressDetails,
-  override val registration: Option[IncorporationRegistrationDetails]
-) extends RegistrationDetails
+case class ChangeOfCircumstanceDetails(
+  changeOfCircumstance: String,
+  deregistrationDetails: Option[DeregistrationDetails] = None
+)
 
-object IncorporationDetails {
+object ChangeOfCircumstanceDetails {
 
-  implicit val format: Format[IncorporationDetails] = Json.format[IncorporationDetails]
+  implicit val format: OFormat[ChangeOfCircumstanceDetails] =
+    Json.format[ChangeOfCircumstanceDetails]
+
 }
