+++ app/uk/gov/hmrc/plasticpackagingtaxregistration/connectors/models/eis/subscription/Subscription.scala
@@ -30,7 +30,9 @@ case class Subscription(
   declaration: Declaration,
   taxObligationStartDate: String,
   last12MonthTotalTonnageAmt: Long,
-  groupPartnershipSubscription: Option[GroupPartnershipSubscription] = None
+  groupPartnershipSubscription: Option[GroupPartnershipSubscription] = None,
+  processingDate: Option[String] = None,
+  changeOfCircumstanceDetails: Option[ChangeOfCircumstanceDetails] = None
 )
 
 object Subscription {
