--- app/uk/gov/hmrc/plasticpackagingtaxregistration/connectors/models/eis/subscriptionDisplay/SubscriptionDisplayResponse.scala
@@ -1,41 +1,0 @@
-/*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.plasticpackagingtaxregistration.connectors.models.eis.subscriptionDisplay
-
-import play.api.libs.json.{Json, OFormat}
-import uk.gov.hmrc.plasticpackagingtaxregistration.connectors.models.eis.subscription._
-import uk.gov.hmrc.plasticpackagingtaxregistration.connectors.models.eis.subscription.group.GroupPartnershipSubscription
-
-case class SubscriptionDisplayResponse(
-  processingDate: String,
-  changeOfCircumstanceDetails: Option[ChangeOfCircumstanceDetails],
-  legalEntityDetails: LegalEntityDetails,
-  principalPlaceOfBusinessDetails: PrincipalPlaceOfBusinessDetails,
-  primaryContactDetails: PrimaryContactDetails,
-  businessCorrespondenceDetails: BusinessCorrespondenceDetails,
-  taxObligationStartDate: String,
-  last12MonthTotalTonnageAmt: BigDecimal,
-  declaration: Declaration,
-  groupOrPartnershipSubscription: Option[GroupPartnershipSubscription]
-)
-
-object SubscriptionDisplayResponse {
-
-  implicit val format: OFormat[SubscriptionDisplayResponse] =
-    Json.format[SubscriptionDisplayResponse]
-
-}
